se hicieron algunas restructuraciones y configurando el controller de customer
