+++ InternetBanking.core/Interfaces/ICustomersService.cs
@@ -1,4 +1,4 @@
-﻿using InternetBanking.core.Dtos;
+﻿using InternetBanking.core.Dtos.Customers;
 using InternetBanking.DataAccess.Models;
 using System;
 using System.Collections.Generic;
